refactor: split model/labelvalue to file
